Changes after rebase; optimiser renamed to optimizer;
