+++ examples/advanced/fedot_based_solutions/graph_model_optimization.py
@@ -8,7 +8,7 @@ import pandas as pd
 from fedot.core.composer.gp_composer.gp_composer import PipelineComposerRequirements
 from fedot.core.dag.verification_rules import has_no_cycle, has_no_self_cycled_nodes
 from fedot.core.optimisers.adapters import DirectAdapter
-from fedot.core.optimisers.gp_comp.gp_optimiser import (
+from fedot.core.optimisers.gp_comp.gp_optimizer import (
     EvoGraphOptimizer,
     GeneticSchemeTypesEnum,
     GPGraphOptimiserParameters
